dentiapro/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.9
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.10
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2024.2
requests==2.32.3
sendgrid==6.10.0
sentry-sdk==1.39.1
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
stripe==6.7.0
twilio==8.10.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
